Break out utility functions into `util.ts` and `util.test.ts`.
Many of these methods are going to be useful for other HTTP transports, so
I've moved them out into their own utility file.

Unfortunately, this commit does actually contain changes other than just a
re-shuffling of methods, though the changes are purely commentary, some
minor formatting and a couple no-op changes to accommodate the function's
new homes.

Sadly, I didn't realize I hadn't staged and committed this file until I did a
fair bit of writing. :facepalm:
